AndroidManifest.xml generation for modules (optional, wip)

+++ code_setup.sh
@@ -179,6 +179,17 @@ function deployFile() {
 		OVER_WRITE=$3;
 	fi
 	echo "--------------------------------------------------------------------------------";
+	if [[ $SRC_FILE_PATH == *.MT.sh ]]; then
+		echo "> Running file '$SRC_FILE_PATH'...";
+		./"$SRC_FILE_PATH";
+		local RESULT=$?;
+		if [[ ${RESULT} -ne 0 ]]; then
+			echo "> Error while running file '$SRC_FILE_PATH'!";
+			exit ${RESULT};
+		fi
+		echo "> Running file '$SRC_FILE_PATH'... DONE";
+		exit ${RESULT};
+	fi
 	if [[ "$OVER_WRITE" == true ]]; then
 		if [[ -f "${DEST_FILE_PATH}" ]]; then
 			rm ${DEST_FILE_PATH};
@@ -324,8 +335,8 @@ done
 echo "> Deploying optional shared files... DONE";
 echo "--------------------------------------------------------------------------------";
 
+echo "--------------------------------------------------------------------------------";
 if [[ $PROJECT_NAME == "mtransit-for-android" ]]; then
-  echo "--------------------------------------------------------------------------------";
   echo "> Deploying main shared files...";
   SRC_DIR_PATH="commons/shared-main";
   for FILENAME in $(ls -a $SRC_DIR_PATH/) ; do
@@ -348,8 +359,31 @@ if [[ $PROJECT_NAME == "mtransit-for-android" ]]; then
     fi
   done
   echo "> Deploying main shared files... DONE";
-  echo "--------------------------------------------------------------------------------";
+else
+  echo "> Deploying modules shared files...";
+  SRC_DIR_PATH="commons/shared-modules";
+  for FILENAME in $(ls -a $SRC_DIR_PATH/) ; do
+    SRC_FILE_PATH=$SRC_DIR_PATH/$FILENAME;
+    if [[ $FILENAME == "." ]] || [[ $FILENAME == ".." ]]; then
+      continue;
+    fi
+    FILENAME_DEST=${FILENAME#"MT"}; # MT+filename used to ignore ".gitignore"
+    DEST_FILE_PATH="$DEST_PATH/$FILENAME_DEST"
+    if [[ -f $SRC_FILE_PATH ]]; then
+      deployFile ${SRC_FILE_PATH} ${DEST_FILE_PATH};
+      checkResult $?;
+    elif [[ -d "$SRC_FILE_PATH" ]]; then
+      deployDirectory ${SRC_FILE_PATH} ${DEST_FILE_PATH};
+      checkResult $?;
+    else #WTF
+      echo "> File to deploy '$FILENAME' ($SRC_FILE_PATH) is neither a directory or a file!";
+      ls -l $FILENAME;
+      exit 1;
+    fi
+  done
+  echo "> Deploying modules shared files... DONE";
 fi
+echo "--------------------------------------------------------------------------------";
 
 echo "--------------------------------------------------------------------------------";
 echo "> Deploying overwritten shared files...";
